refactor: reorganiza estrutura de componentes
