+++ apps/medicamentar-desktop/src/components/Modals/NewMedicine.tsx
@@ -1,0 +1,8 @@
+const NewMedicine = () => {
+return (
+    <></>
+)
+
+};
+
+export default NewMedicine;
